agenda read only working
