--- src/component/data/example.js
@@ -1,19 +1,0 @@
-export const salon = {
-    employe: ["Rausis Justin", "Peer Vincent", "De Blesser Dimitri"],
-    appointments: [
-        {
-            title: 'Coupe femme',
-            startDate: new Date(2018, 5, 25, 9, 35),
-            endDate: new Date(2018, 5, 25, 11, 30),
-            id: 0,
-            location: 'Rausis Justin',
-        },
-        {
-                title: 'Coupe homme',
-                startDate: new Date(2018, 5, 25, 16, 35),
-                endDate: new Date(2018, 5, 25, 17, 30),
-                id: 1,
-                location: 'Peer Vincent',
-        }
-    ]
-}
